+++ tool_service/tool_node/assets/configuration.json
@@ -1,0 +1,6 @@
+{
+  "name": "RenewInstance",
+  "amap_weather" : {
+    "key" : "test"
+  }
+}
